+++ src/krm3/management/commands/check_images.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python3
 
-import os
-import sys
 import re
 from pathlib import Path
 from collections import defaultdict