+++ src/krm3/timesheet/dto.py
@@ -1,6 +1,7 @@
 """Non-model domain data-transfer-objects for the timesheet."""
 
 import datetime
+import typing
 from typing import Self
 
 from krm3.core.models.auth import Resource, User
@@ -9,9 +10,12 @@ from krm3.core.models.timesheets import TimeEntry, TimeEntryQuerySet
 from krm3.utils.dates import KrmCalendar
 from constance import config
 
+if typing.TYPE_CHECKING:
+    from krm3.config.fragments.constance import ConstanceTyping
+
 
 class TimesheetDTO:
-    def __init__(self, requested_by: User) -> None:
+    def __init__(self, requested_by: User | None = None) -> None:
         self.tasks = TaskQuerySet().none()
         self.time_entries = TimeEntryQuerySet().none()
         self._requested_by = requested_by
@@ -21,14 +25,11 @@ class TimesheetDTO:
         self.bank_hours = 0.0
 
     def fetch(self, resource: Resource, start_date: datetime.date, end_date: datetime.date) -> Self:
-        self.tasks = (
-            Task.objects.filter_acl(self._requested_by)  # pyright: ignore[reportAttributeAccessIssue]
-            .active_between(start_date, end_date)
-            .assigned_to(resource=resource)
-        )
-        self.time_entries = TimeEntry.objects.filter_acl(self._requested_by).filter(  # pyright: ignore[reportAttributeAccessIssue]
-            resource=resource, date__range=(start_date, end_date)
-        )
+        """Fetch the resource timesheet for a specific date interval."""
+        task_qs = Task.objects.filter_acl(self._requested_by) if self._requested_by else Task.objects
+        self.tasks = task_qs.active_between(start_date, end_date).assigned_to(resource=resource)
+        te_qs = TimeEntry.objects.filter_acl(self._requested_by) if self._requested_by else TimeEntry.objects
+        self.time_entries = te_qs.filter(resource=resource, date__range=(start_date, end_date))
         calendar = KrmCalendar()
 
         self.days = calendar.iter_dates(start_date, end_date)
@@ -36,12 +37,17 @@ class TimesheetDTO:
 
         self.schedule = resource.get_schedule(start_date, end_date)
 
-        self.timesheet_colors['less_than_schedule_color_bright_theme'] = config.LESS_THAN_SCHEDULE_COLOR_BRIGHT_THEME
-        self.timesheet_colors['exact_schedule_color_bright_theme'] = config.EXACT_SCHEDULE_COLOR_BRIGHT_THEME
-        self.timesheet_colors['more_than_schedule_color_bright_theme'] = config.MORE_THAN_SCHEDULE_COLOR_BRIGHT_THEME
-        self.timesheet_colors['less_than_schedule_color_dark_theme'] = config.LESS_THAN_SCHEDULE_COLOR_DARK_THEME
-        self.timesheet_colors['exact_schedule_color_dark_theme'] = config.EXACT_SCHEDULE_COLOR_DARK_THEME
-        self.timesheet_colors['more_than_schedule_color_dark_theme'] = config.MORE_THAN_SCHEDULE_COLOR_DARK_THEME
+        conf: ConstanceTyping = config
+        self.timesheet_colors.update(
+            {
+                'less_than_schedule_color_bright_theme': conf.LESS_THAN_SCHEDULE_COLOR_BRIGHT_THEME,
+                'exact_schedule_color_bright_theme': conf.EXACT_SCHEDULE_COLOR_BRIGHT_THEME,
+                'more_than_schedule_color_bright_theme': conf.MORE_THAN_SCHEDULE_COLOR_BRIGHT_THEME,
+                'less_than_schedule_color_dark_theme': conf.LESS_THAN_SCHEDULE_COLOR_DARK_THEME,
+                'exact_schedule_color_dark_theme': conf.EXACT_SCHEDULE_COLOR_DARK_THEME,
+                'more_than_schedule_color_dark_theme': conf.MORE_THAN_SCHEDULE_COLOR_DARK_THEME,
+            }
+        )
 
         self.bank_hours = resource.get_bank_hours_balance()