+++ tests/api/timesheet/test_timesheet_model_viewset.py
@@ -1,18 +1,17 @@
 import typing
 
 import pytest
-
-from testutils.factories import (
-    ResourceFactory,
-    UserFactory,
-    TimesheetSubmissionFactory,
-    GroupFactory,
-)
 from rest_framework import status
 from rest_framework.reverse import reverse
+from testutils.factories import (
+    GroupFactory,
+    ResourceFactory,
+    TimesheetSubmissionFactory,
+    UserFactory,
+)
+from testutils.permissions import add_group_permissions, add_permissions
 
 from krm3.core.models import TimesheetSubmission
-from testutils.permissions import add_permissions, add_group_permissions
 
 if typing.TYPE_CHECKING:
     from krm3.core.models import Resource
@@ -158,3 +157,17 @@ class TestTimesheetSubmissionModelAPIListView:
             f'DETAIL:  Key (period, resource_id)=([2024-01-06,2024-01-16), {resource.pk})'
             f' conflicts with existing key (period, resource_id)=([2024-01-01,2024-01-07), {resource.pk}).'
         }
+
+    def test_resubmit_open_timesheet(self, api_client, regular_user):
+        resource: Resource = ResourceFactory(user=regular_user)
+        # create and existing TimesheetSubmission with the same resource and overlapping period
+        ts: TimesheetSubmission = TimesheetSubmissionFactory(
+            resource=resource, period=('2024-01-01', '2024-01-08'), closed=False
+        )
+        assert ts.timesheet is None
+        response = api_client(user=regular_user).post(
+            self.url(), data={'resource': resource.pk, 'period': ('2024-01-01', '2024-01-07')}, format='json'
+        )
+        assert response.status_code == status.HTTP_201_CREATED
+        ts = TimesheetSubmission.objects.get(pk=response.data['id'])
+        assert ts.timesheet is not None